Fix format string assertion in select net dialog.

+++ pcbnew/dialogs/dialog_select_net_from_list.cpp
@@ -2,7 +2,7 @@
  * This program source code file is part of KiCad, a free EDA CAD application.
  *
  * Copyright (C) 2016 Jean-Pierre Charras, jp.charras at wanadoo.fr
- * Copyright (C) 1992-2016 KiCad Developers, see AUTHORS.txt for contributors.
+ * Copyright (C) 1992-2017 KiCad Developers, see AUTHORS.txt for contributors.
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -154,7 +154,7 @@ void DIALOG_SELECT_NET_FROM_LIST::buildNetsList()
 
         if( netcode )
         {
-            txt.Printf( wxT( "%u" ), net->m_PadInNetList.size() );
+            txt.Printf( wxT( "%u" ), (unsigned) net->m_PadInNetList.size() );
             m_netsListGrid->SetCellValue( row_idx, COL_NETINFO, txt );
         }
         else    // For the net 0 (unconnected pads), the pad count is not known
